Move namespace related enums to _enums.py
